\name{PKH26data}

\alias{PKH26data}

\docType{data}

\title{ genFitting: example data with PKH26 stain }

\description{
  example data with PKH26 dye.
  Two samples: NONPROL (parent population) and PROL (proliferating population).
}

\usage{data(PKH26data)}

\format{
  The format is an object of class \code{\link{flowSet}} with 2 \code{\link{flowFrame}}
}

\keyword{datasets}
